# Receive messages perf improvements