--- a/src/status_im/transport/message/core.cljs
+++ b/src/status_im/transport/message/core.cljs
@@ -26,13 +26,19 @@
   dedup-id is passed by status-go and is used to deduplicate messages at that layer.
   Once a message has been successfuly processed, that id needs to be sent back
   in order to stop receiving that message"
-  [cofx now-in-s filter-chat-id message]
+  [cofx now-in-s filter-chat-id message-js]
   (let [blocked-contacts (get-in cofx [:db :contacts/blocked] #{})
-        {{:keys [payload timestamp]} :message
-         metadata :metadata
-         raw-payload :raw-payload} (add-raw-payload message)
+        payload (.-payload message-js)
+        timestamp (.-timestamp (.-message message-js))
+        metadata-js (.-metadata message-js)
+        metadata {:author {:publicKey (.-publicKey (.-author metadata-js))
+                           :alias (.-alias (.-author metadata-js))
+                           :identicon (.-identicon (.-author metadata-js))}
+                  :dedupId (.-dedupId metadata-js)
+                  :encryptionId (.-encryptionId metadata-js)
+                  :messageId (.-messageId metadata-js)}
+        raw-payload  {:raw-payload message-js}
         status-message (-> payload
-                           ethereum/hex-to-utf8
                            transit/deserialize)
         sig (-> metadata :author :publicKey)]
     (when (and sig
@@ -70,21 +76,24 @@
       (apply fx/merge cofx receive-message-fxs))
     (log/error "Something went wrong" error messages)))
 
-(fx/defn receive-messages [cofx event]
+(fx/defn receive-messages [cofx event-js]
   (let [fxs (keep
-             (fn [{:keys [chat messages error]}]
-               (when (seq messages)
-                 (receive-whisper-messages
-                  error
-                  messages
+             (fn [message-specs]
+               (let [chat (.-chat message-specs)
+                     messages (.-messages message-specs)
+                     error (.-error message-specs)]
+                 (when (seq messages)
+                   (receive-whisper-messages
+                    error
+                    messages
                   ;; For discovery and negotiated filters we don't
                   ;; set a chatID, and we use the signature of the message
                   ;; to indicate which chat it is for
-                  (if (or (:discovery chat)
-                          (:negotiated chat))
-                    nil
-                    (:chatId chat)))))
-             (:messages event))]
+                    (if (or (.-discovery chat)
+                            (.-negotiated chat))
+                      nil
+                      (.-chatId chat))))))
+             (.-messages event-js))]
     (apply fx/merge cofx fxs)))
 
 (fx/defn remove-hash