… externs.js
… src/status_im/ui/screens/navigation.cljs
… status-go-version.json