@status-im-auto 2019-10-22
### Jenkins Builds
| :grey_question: | Commit | :hash: | Finished (UTC) | Duration | Platform | Result |
|-|-|-|-|-|-|-|
| :x: | 82796241 | [#1](https://ci.status.im/job/status-react/job/prs/job/linux/job/PR-9263/1/) | 2019-10-22 14:37:08 | ~29 sec | `linux` | [:page_facing_up:`log`](https://ci.status.im/job/status-react/job/prs/job/linux/job/PR-9263/1/consoleText) |