@rasom 2019-10-23
@cammellos 
> Both events should be handle in 16ms for keeping the ui responsive.
Investigation showed that neither or them is close to that.

on any device?