@rasom 2019-10-23
Do we have any specific numbers how faster conversion became after these changes?