@cammellos 2019-10-23
@Serhy benchmarked it here (same code when it was benchmarked as this PR https://github.com/status-im/status-react/pull/9230#issuecomment-544466568 , later changes have not been included as I was testing stuff)