@cammellos 2019-10-23
@rasom 

> on any device?

I didn't get anywhere close to that on any device ( I don't have top of the range devices, but I have an Huawei mediapad which should have decent performance), for what is worth in desktop running emulator (with the built apk) I also was not able to be under 16ms for the second part of the event, although not sure that's reliable.