@churik 2019-10-23
@cammellos no, it is not false failure - it is an issue, however, I can't reproduce it manually.
Basically sometimes list of assets is not loaded and instead of all your tokens / collectibles you will see only  ETH.

Issue is pretty serious, but I didn't see it when tested manually